Moved SqlRow from _abc
